+++ delivery-service/src/main/java/ru/liga/service/api/DeliveryService.java
@@ -7,6 +7,5 @@ import ru.liga.dto.Status;
 public interface DeliveryService {
     DeliveryAllDto getAllDelivers(long id);
     Status create(String phone);
-    void updateStatusById(long orderId,long courerId);
     String dropFeign(Long id);
  }
