+++ delivery-service/src/main/java/ru/liga/service/api/RabbitListenerService.java
@@ -1,0 +1,8 @@
+package ru.liga.service.api;
+
+import org.springframework.stereotype.Service;
+
+
+public interface RabbitListenerService {
+    void updateStatusById(long orderId,long courerId);
+}
